refactor(geo): move the client and service into Appwrite\Geo
Geo lookups are their own domain, not a Locale concern. GeoRecord stays
in Appwrite\Locale since it renders locale-aware names.
