+++ app/init/resources/request.php
@@ -14,9 +14,9 @@ use Appwrite\Event\Realtime;
 use Appwrite\Event\Webhook;
 use Appwrite\Extend\Exception;
 use Appwrite\Functions\EventProcessor;
+use Appwrite\Geo\Client as GeoClient;
+use Appwrite\Geo\Geo;
 use Appwrite\GraphQL\Schema;
-use Appwrite\Locale\Geo;
-use Appwrite\Locale\Geo\Client as GeoClient;
 use Appwrite\Locking\Lock;
 use Appwrite\Network\Cors;
 use Appwrite\Network\Platform;
